Separate testing crate content into model, loader, and execition
