+++ crates/testing/src/execution/introspection_tests.js
