+++ crates/testing/src/loader/mod.rs
@@ -1,0 +1,13 @@
+// Copyright Exograph, Inc. All rights reserved.
+//
+// Use of this software is governed by the Business Source License
+// included in the LICENSE file at the root of this repository.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0.
+
+mod integration_test;
+mod integration_tests;
+
+pub(crate) use integration_tests::IntegrationTests;
